feat(errors): translate error messages to French
Translate all centralized error messages (errors.ts, upload route, DTOs)
to French and update corresponding test assertions to match.
