+++ backend/src/errors/errors.ts
@@ -2,11 +2,11 @@ import { AppError } from "./AppError";
 
 export const Errors = {
     notFound: (entity: string) =>
-        new AppError("NOT_FOUND", `${entity} not found`),
+        new AppError("NOT_FOUND", `${entity} introuvable`),
 
     alreadyExists: (entity: string) =>
-        new AppError("ALREADY_EXISTS", `${entity} already exists`),
-    
+        new AppError("ALREADY_EXISTS", `${entity} existe déjà`),
+
     unauthorized: () =>
-        new AppError("UNAUTHORIZED", "You are not authorized"),
+        new AppError("UNAUTHORIZED", "Vous n'êtes pas autorisé"),
 };
